requirements.txt
altgraph==0.17.4
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
Flask==3.0.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pycairo==1.20.1
PyGObject==3.42.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.3.0
pyparsing==2.4.7
pystray==0.19.5
pytest==8.3.3
python-apt==2.4.0+ubuntu4
python-engineio==4.10.1
python-socketio==5.11.4
python-xlib==0.33
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
ssh-import-id==5.11
systemd-python==234
tomli==2.0.2
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
zipp==1.0.0
